Minimap Refactoring + Race/Orientation Store created
Seperated the minimap from being in the index to being in a Minimap.tsx, added Race/Orientation stores to hold information like areas discovered/locations found. Passed the mode state to the minimap so that if the user is in the amazing race game that it displays the areas discovered from the race store not the orientation store and vice versa.